+++ client/allocrunnerdeprecated/taskrunner/task_runner.go
@@ -21,8 +21,8 @@ import (
 	"github.com/hashicorp/go-multierror"
 	version "github.com/hashicorp/go-version"
 	"github.com/hashicorp/nomad/client/allocdir"
-	"github.com/hashicorp/nomad/client/allocrunner/getter"
-	"github.com/hashicorp/nomad/client/allocrunner/taskrunner/restarts"
+	"github.com/hashicorp/nomad/client/allocrunnerv2/taskrunner/getter"
+	"github.com/hashicorp/nomad/client/allocrunnerv2/taskrunner/restarts"
 	"github.com/hashicorp/nomad/client/config"
 	consulApi "github.com/hashicorp/nomad/client/consul"
 	"github.com/hashicorp/nomad/client/driver"
